Refactor cart slice, initialize profile slice
