+++ src/modal/Auth/auth.module.css
@@ -4,8 +4,8 @@
     right: 0;
     left: 0;
     bottom: 0;
-    max-width: 40vw;
-    max-height: 60vh;
+    max-width: 500px;
+    max-height: max-content;
     background-color: white !important;
     border: 2px solid #000;
     padding: 50px;
